Support subfolders in local song storage (#42)
This change allows subfolders in the local song storage. It changes the song list to work recursively and also allows to use e.g. "Existing Folder/Song name" when saving a new song. Otherwise this does not really advertise the subfolder possibility, existing subfolders can be used, but there is no special support for them in the UI.

- Support subfolders in local song storage
- Support saving into existing local subfolders